refactor(listen): split upcoming show card views
